requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-Mail==0.10.0
Flask-Caching==2.3.0
Flask-CORS==4.0.1
SQLAlchemy==2.0.30
Werkzeug==3.0.3
celery==5.4.0
redis==5.0.7
kombu==5.3.7
python-dotenv==1.0.1
marshmallow==3.21.3
Pillow==10.3.0
reportlab==4.2.0
